Move checks of invalid required behaviors before the preview to its own function
